Improved assimp loading Shader now gives proper error if needed Added image initialisation Added filesystem reading of directory

+++ VrLib/Model.h
@@ -67,7 +67,7 @@ namespace vrlib
 
 		virtual std::vector<glm::vec3> getVertices(int amount) const = 0;
 		virtual std::vector<glm::vec3> getTriangles() const = 0;
-		virtual std::pair<std::vector<unsigned short>, std::vector<glm::vec3>> getIndexedTriangles() const { throw "not implemented"; };
+		virtual std::pair<std::vector<unsigned int>, std::vector<glm::vec3>> getIndexedTriangles() const { throw "not implemented"; };
 		virtual void draw(const std::function<void(const glm::mat4&)> &modelviewMatrixCallback, const std::function<void(const Material&)> &materialCallback = nullptr) = 0;
 
 		virtual ModelInstance* getInstance() = 0;
